+++ app/src/main/java/algonquin/cst2335/hhsrandroidproject1/oct/OCTOpenHelper.java
@@ -4,9 +4,6 @@ import android.content.Context;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
 
-import androidx.annotation.Nullable;
-
-import java.io.InputStream;
 
 public class OCTOpenHelper extends SQLiteOpenHelper {
     public static final String name = "OCTDatabase";
