+++ app/src/main/java/algonquin/cst2335/hhsrandroidproject1/oct/RouteFragment.java
@@ -1,0 +1,143 @@
+package algonquin.cst2335.hhsrandroidproject1.oct;
+import android.os.Bundle;
+import android.util.Log;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.TextView;
+
+import androidx.fragment.app.Fragment;
+
+import org.json.JSONArray;
+import org.json.JSONException;
+import org.json.JSONObject;
+
+import java.io.BufferedInputStream;
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.net.HttpURLConnection;
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.nio.charset.StandardCharsets;
+import java.util.ArrayList;
+import java.util.concurrent.Executor;
+import java.util.concurrent.Executors;
+import java.util.stream.Collectors;
+
+import algonquin.cst2335.hhsrandroidproject1.R;
+
+public class RouteFragment extends Fragment{
+    private String stringURL;
+    String routeDirection_routeLabel;
+    String direction;
+    String routeNumber;
+    ArrayList<Trip> tripArrayList = new ArrayList<>();
+    @Override
+    public View onCreateView( LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+       View routeLayout = inflater.inflate(R.layout.route_detail_layout,container,false);
+
+        TextView titleText = routeLayout.findViewById(R.id.routeHeader);
+        TextView stopDescriptionText = routeLayout.findViewById(R.id.routeDetailInfo);
+        String stationNumber = getActivity().getIntent().getStringExtra("StationNumber");//"3017";
+        String routeNo = getActivity().getIntent().getStringExtra("RouteNo");//"75";
+
+        Executor newThread = Executors.newSingleThreadExecutor();
+        // create a stop list in the stop recyclerView list
+        newThread.execute(() -> {//this runs on another thread
+            try {
+                stringURL = "https://api.octranspo1.com/v2.0/GetNextTripsForStop?appID=223eb5c3&&apiKey=ab27db5b435b8c8819ffb8095328e775&stopNo="
+                        + stationNumber + " &routeNo=" + routeNo + "&format=json";
+                URL url = new URL(stringURL);
+                HttpURLConnection urlConnection = (HttpURLConnection) url.openConnection();
+                InputStream in = new BufferedInputStream(urlConnection.getInputStream());
+                String text = (new BufferedReader(//To read this data in Android, you have to convert the inputStream from the server into a Java String object
+                        new InputStreamReader(in, StandardCharsets.UTF_8)))
+                        .lines()
+                        .collect(Collectors.joining("\n"));
+                //convert string to JSON object
+                JSONObject theDocument = new JSONObject(text);
+                JSONObject getNextTripsForStopResult = theDocument.getJSONObject("GetNextTripsForStopResult");
+                String stopNo = getNextTripsForStopResult.getString("StopNo");
+                String error = getNextTripsForStopResult.getString("Error");
+                String stopLabel = getNextTripsForStopResult.getString("StopLabel");
+                //titleText.setText("Information of the Route: " );
+                //stopDescriptionText.setText(stopLabel +error);
+                JSONObject routesObj = getNextTripsForStopResult.getJSONObject("Route");
+                JSONArray routeDirectionArray = routesObj.getJSONArray("RouteDirection");
+                for (int i = 0; i < routeDirectionArray.length(); i++) {
+                    JSONObject o = routeDirectionArray.getJSONObject(i);
+                    routeNumber = o.getString("RouteNo");
+                    routeDirection_routeLabel = o.getString("RouteLabel");
+                    direction = o.getString("Direction");
+                    JSONObject trips = o.getJSONObject("Trips");
+                    JSONArray tripArray = trips.getJSONArray("Trip");
+                    for (int counter = 0; counter < tripArray.length(); counter++){
+                        JSONObject object = tripArray.getJSONObject(counter);
+                        Trip iterator = new Trip();
+                        iterator.trip_longitude = object.getString("Longitude");
+                        iterator.trip_latitude = object.getString("Latitude");
+                        iterator.trip_gpsSpeed = object.getString("GPSSpeed");
+                        iterator.trip_destination = object.getString("TripDestination");
+                        iterator.trip_startTime = object.getString("TripStartTime");
+                        iterator.trip_adjustedScheduleTime = object.getString("AdjustedScheduleTime");
+                        iterator.Trip_adjustmentAge = object.getString("AdjustmentAge");
+                        iterator.trip_lastTripOfSchedule = object.getBoolean("LastTripOfSchedule");
+                        iterator.trip_busType = object.getString("BusType");
+                        tripArrayList.add(iterator);
+                    }
+                }
+                getActivity().runOnUiThread(() -> {
+                    titleText.setText("Information of the Route: " + stopNo + stopLabel + error);
+                    stopDescriptionText.setText(getTripsText(tripArrayList));
+                });
+
+            } catch (MalformedURLException e) {
+                e.printStackTrace();
+            } catch(JSONException joe) {
+                getActivity().runOnUiThread(() -> {
+                    titleText.setText("Can't retrieve information of trips for Route: " + routeNo + " Stop:"+ stationNumber);
+                    stopDescriptionText.setText("");
+                });
+            } catch (IOException ioe) {
+                Log.e("Connection error: ", ioe.getMessage());
+            }
+        });
+        return routeLayout;
+    }
+
+    public String getTripsText(ArrayList<Trip> tripArrayList) {
+        StringBuilder sb = new StringBuilder();
+        int count = 0;
+        String dest = "";
+        for (Trip it : tripArrayList) {
+            if(!dest.equals(it.trip_destination)) {
+                sb.append("\nDestination: ");
+                sb.append(it.trip_destination);
+                sb.append("\n");
+                dest = it.trip_destination;
+            }
+            sb.append("  Trip");
+            sb.append(++count);
+            sb.append(":\n    Longitude:");
+            sb.append(it.trip_longitude);
+            sb.append("\n    latitude:");
+            sb.append(it.trip_latitude);
+            sb.append("\n");
+        }
+        return sb.toString();
+    }
+    private class Trip {
+        String trip_longitude ;
+        String trip_latitude ;
+        String trip_gpsSpeed ;
+        String trip_destination ;
+        String trip_startTime ;
+        String trip_adjustedScheduleTime ;
+        String Trip_adjustmentAge ;
+        boolean trip_lastTripOfSchedule ;
+        String trip_busType;
+
+    }
+}
