+++ app/src/main/res/layout/route_detail_layout.xml
@@ -6,10 +6,11 @@
     android:layout_height="match_parent">
 
     <TextView
-        android:id="@+id/route_header"
-        android:layout_width="257dp"
-        android:layout_height="43dp"
+        android:id="@+id/routeHeader"
+        android:layout_width="279dp"
+        android:layout_height="44dp"
         android:layout_marginTop="16dp"
+        android:fontFamily="@font/archivo_black"
         android:text="title goes here"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
@@ -17,12 +18,13 @@
 
     <TextView
         android:id="@+id/routeDetailInfo"
-        android:layout_width="283dp"
-        android:layout_height="584dp"
-        android:layout_marginTop="8dp"
+        android:layout_width="279dp"
+        android:layout_height="671dp"
+        android:fontFamily="@font/baloo_bhai"
         android:text="route detail goes here"
+        android:textSize="15sp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/route_header" />
+        app:layout_constraintTop_toBottomOf="@+id/routeHeader" />
 </androidx.constraintlayout.widget.ConstraintLayout>
